feat: add random link to login page
