+++ src/components/cities/cities.tsx
@@ -1,25 +1,16 @@
 
 import LocationItemLink from '../../components/location-item-link/location-item-link';
-import {MouseEvent} from 'react';
-import {useAppDispatch, useAppSelector} from '../../hooks';
-import {changeCity, getOffers} from '../../store/action';
+import {useAppSelector} from '../../hooks';
 
 type CitiesProps = {
   cities: string[];
 }
 
 function Cities({cities}: CitiesProps): JSX.Element {
-  const dispatch = useAppDispatch();
   const activeCity = useAppSelector((state) => state.city);
-  const handleListCitiesItemClick = (evt: MouseEvent<HTMLElement>) => {
-    if (evt.currentTarget.textContent) {
-      dispatch(changeCity(evt.currentTarget.textContent));
-      dispatch(getOffers());
-    }
-  };
   const citiesUl = cities &&
     cities.map((city)=>
-      <li className='locations__item' key={city} onClick={handleListCitiesItemClick}><LocationItemLink text={city} isTab isActive={city === activeCity}/></li>
+      <li className='locations__item' key={city}><LocationItemLink text={city} isTab isActive={city === activeCity}/></li>
     );
   return (
     <>
